Cleans up PS2Keyboard.cpp into c file.
Also adds necessary functions to the scan loop file to output PS2 bytes.
