+++ Scan/PS2/PS2Keyboard.c
@@ -1,0 +1,101 @@
+/*
+  PS2Keyboard.cpp - PS2Keyboard library
+  Copyright (c) 2007 Free Software Foundation.  All right reserved.
+  Written by Christian Weichel <info@32leaves.net>
+
+  ** Mostly rewritten Paul Stoffregen <paul@pjrc.com> 2010, 2011
+  ** Modified for use beginning with Arduino 13 by L. Abraham Smith, <n3bah@microcompdesign.com> * 
+  ** Modified for easy interrup pin assignement on method begin(datapin,irq_pin). Cuningan <cuninganreset@gmail.com> **
+
+  for more information you can read the original wiki in arduino.cc
+  at http://www.arduino.cc/playground/Main/PS2Keyboard
+  or http://www.pjrc.com/teensy/td_libs_PS2Keyboard.html
+
+  Version 2.4 (March 2013)
+  - Support Teensy 3.0, Arduino Due, Arduino Leonardo & other boards
+  - French keyboard layout, David Chochoi, tchoyyfr at yahoo dot fr
+
+  Version 2.3 (October 2011)
+  - Minor bugs fixed
+
+  Version 2.2 (August 2011)
+  - Support non-US keyboards - thanks to Rainer Bruch for a German keyboard :)
+
+  Version 2.1 (May 2011)
+  - timeout to recover from misaligned input
+  - compatibility with Arduino "new-extension" branch
+  - TODO: send function, proposed by Scott Penrose, scooterda at me dot com
+
+  Version 2.0 (June 2010)
+  - Buffering added, many scan codes can be captured without data loss
+    if your sketch is busy doing other work
+  - Shift keys supported, completely rewritten scan code to ascii
+  - Slow linear search replaced with fast indexed table lookups
+  - Support for Teensy, Arduino Mega, and Sanguino added
+
+  This library is free software; you can redistribute it and/or
+  modify it under the terms of the GNU Lesser General Public
+  License as published by the Free Software Foundation; either
+  version 2.1 of the License, or (at your option) any later version.
+
+  This library is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+  Lesser General Public License for more details.
+
+  You should have received a copy of the GNU Lesser General Public
+  License along with this library; if not, write to the Free Software
+  Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+*/
+
+#include "PS2Keyboard.h"
+
+#define BUFFER_SIZE 45
+static volatile uint8_t buffer[BUFFER_SIZE];
+static volatile uint8_t head, tail;
+
+// The ISR for the external interrupt
+void ps2interrupt(void)
+{
+	static uint8_t bitcount=0;
+	static uint8_t incoming=0;
+	static uint32_t prev_ms=0;
+	uint32_t now_ms;
+	uint8_t n, val;
+
+	val = ps2data_read();
+	now_ms = millis();
+	if (now_ms - prev_ms > 250) {
+		bitcount = 0;
+		incoming = 0;
+	}
+	prev_ms = now_ms;
+	n = bitcount - 1;
+	if (n <= 7) {
+		incoming |= (val << n);
+	}
+	bitcount++;
+	if (bitcount == 11) {
+		uint8_t i = head + 1;
+		if (i >= BUFFER_SIZE) i = 0;
+		if (i != tail) {
+			buffer[i] = incoming;
+			head = i;
+		}
+		bitcount = 0;
+		incoming = 0;
+	}
+}
+
+uint8_t get_scan_code(void)
+{
+	uint8_t c, i;
+
+	i = tail;
+	if (i == head) return 0;
+	i++;
+	if (i >= BUFFER_SIZE) i = 0;
+	c = buffer[i];
+	tail = i;
+	return c;
+}
